split out batch tests
